refactor(git): split diff target model
